PIFu/requirements.txt
cycler==0.10.0
decorator==4.4.1
imageio==2.8.0
kiwisolver==1.1.0
matplotlib==3.1.3
networkx==2.4
numpy==1.21.6
opencv-python==4.1.2.30
pathlib==1.0.1
Pillow==9.1.0
PyOpenGL==3.1.6
pyparsing==2.4.6
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
torch==1.4.0
torchvision==0.5.0
git+https://github.com/huggingface/diffusers
trimesh==3.5.23
tqdm==4.64.0
cycler==0.10.0
decorator==4.4.1
imageio==2.8.0
kiwisolver==1.1.0
matplotlib==3.1.3
networkx==2.4
numpy==1.21.6
opencv-python==4.1.2.30
pathlib==1.0.1
Pillow==9.1.0
PyOpenGL==3.1.6
pyparsing==2.4.6
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
torch==1.4.0
torchvision==0.5.0
trimesh==3.5.23
tqdm==4.64.0
transformers
accelerate
ftfy